requirements.txt
beautifulsoup4==4.9.0
Flask==1.1.2
requests==2.23.0
Pillow==7.1.2
